ticket-service/requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.9
pika==1.3.2
python-dotenv==1.0.0
